% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Amount}
\alias{Amount}
\title{Amount data}
\format{
A data frame with 10,000 rows and 9 variables:
\describe{
  \item{date}{Date in format yyyy-mm-dd}
  \item{week}{Date in format yyyy-'W'ww}
  \item{month}{Date in format yyyy-mm}
  \item{quarter}{Date in format yyyy-'Q'q}
  \item{year}{Date in format yyyy}
  \item{item}{Item ID}
  \item{itemgroup}{Item group ID}
  \item{amount}{Item amount}
  \item{value}{Item value}
}
}
\source{
anonymized real data
}
\usage{
Amount
}
\description{
A dataset containing 23 items and their amounts over 3 years of data.
}
\keyword{datasets}
