% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCXYZanalysis.R
\docType{methods}
\name{compare}
\alias{compare}
\title{Compares two S4 objects}
\usage{
compare(object1, object2, ...)
}
\arguments{
\item{object1}{First S4 object.}

\item{object2}{Second S4 object.}

\item{...}{Further comparison parameters.}
}
\value{
Comparison result.
}
\description{
Compares two S4 objects.
}
\examples{
data("Amount")
data1 = Amount[sample(1:nrow(Amount), 1000),]
data2 = Amount[sample(1:nrow(Amount), 1000),]
abcxyzData1 = computeABCXYZAnalysis(data1, value = "value", item = "item", timestamp = "date")
abcxyzData2 = computeABCXYZAnalysis(data2, value = "value", item = "item", timestamp = "date")
comparison = compare(abcxyzData1, abcxyzData2)
}
\seealso{
\code{\link[=ABCXYZComparison-class]{compare}}
}
\author{
Leon Binder \email{leon.binder@th-deg.de}

Bernhard Bauer \email{bernhard.bauer@th-deg.de}

Michael Scholz \email{michael.scholz@th-deg.de}
}
\keyword{methods}
