% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performer.R
\name{computeUnderperformer}
\alias{computeUnderperformer}
\title{Select underperforming items}
\usage{
computeUnderperformer(
  data,
  value,
  group,
  timestamp,
  timestampFormat = c("day", "week", "month", "quater", "year"),
  currentTime,
  thresholdValue = 0,
  thresholdTime = 90
)
}
\arguments{
\item{data}{Dataframe containing item stock data.}

\item{value}{Name of the column variable containing the stock values.}

\item{group}{Name(s) of the column(s) that are used to group stock data. These columns are usually the item ID or item name to group
stock data by items.}

\item{timestamp}{Name of the column including the timestamp. This column should be in Date, POSIX , YY-mm, YYYY-'W'ww, YYYY-mm, YYYY-'Q'q or YYYY format.}

\item{timestampFormat}{Declares in which format the timestamp comes in (i.e., "day", "week", "month", "quarter", "year")}

\item{currentTime}{Qualifying date for the value variable. Date must exist in data and have the same format as timestamp-variable.}

\item{thresholdValue}{Name of the colum variable containing the items' stock threshold value or the threshold value used in this analysis
for all items.}

\item{thresholdTime}{Time for which the value shouldn't exceed the threshold value. Number declares the time in the format of timestampFormat}
}
\value{
Returns a data frame listing all underperforming items, the date their stock was the last time over the threshold (lastover),
the duration in days since the stock is under the threshold (toolowindays), the average difference between the stock and the threshold
(meandiff) and the count of switched between over- and underperformance (moves).
}
\description{
Selects items with a value lower than a given threshold for a specified time period.
}
\examples{
data("Stocks")
underperformer = computeUnderperformer(data=Stocks,
                                       value = "stock",
                                       group = "item",
                                       timestamp = "date",
                                       timestampFormat = "day",
                                       currentTime = "2019-07-27",
                                       thresholdValue = "minstock",
                                       thresholdTime = 90)
}
\seealso{
\code{\link{computeOverperformer}} \code{\link{computeConstants}}
}
\author{
Leon Binder \email{leon.binder@th-deg.de}

Bernhard Bauer \email{bernhard.bauer@th-deg.de}

Michael Scholz \email{michael.scholz@th-deg.de}
}
