% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{matmanDemo}
\alias{matmanDemo}
\title{Launches a demo app}
\usage{
matmanDemo()
}
\description{
Launches a shiny app that demonstrates how to use the functions provides by package matman.
}
\author{
Leon Binder \email{leon.binder@th-deg.de}

Bernhard Bauer \email{bernhard.bauer@th-deg.de}

Michael Scholz \email{michael.scholz@th-deg.de}
}
