% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forecast.R
\docType{methods}
\name{show,Forecast-method}
\alias{show,Forecast-method}
\title{Shows a Forecast object}
\usage{
\S4method{show}{Forecast}(object)
}
\arguments{
\item{object}{The \code{Forecast} object}
}
\description{
Shows the predicted data of a Forecast object. If the Forecast object was created using keepPreviousData = TRUE, also the training data
are shown
}
\examples{
data("Amount")
prediction = predictValue(data = Amount,
    value = "amount",
    item = "item",
    timestamp = "date",
    temporalAggregation = "week",
    timeUnitsAhead = 3)
prediction
}
\seealso{
\code{\link[=Forecast-class]{Forecast}}
}
\author{
Leon Binder \email{leon.binder@th-deg.de}

Bernhard Bauer \email{bernhard.bauer@th-deg.de}

Michael Scholz \email{michael.scholz@th-deg.de}
}
\keyword{methods}
