% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Stocks}
\alias{Stocks}
\title{Stock data}
\format{
A data frame with 1,610 rows and 5 variables:
\describe{
  \item{date}{Date in format yyyy-mm-dd}
  \item{item}{Item ID}
  \item{stock}{Item stock value}
  \item{minstock}{Minimum stock per item}
  \item{reorderlevel}{Stock threshold for triggering item reorders}
}
}
\source{
anonymized real data
}
\usage{
Stocks
}
\description{
A dataset containing 10 items and their stocks over 3 years of data.
}
\keyword{datasets}
