% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCXYZanalysis.R
\docType{methods}
\name{summary}
\alias{summary}
\title{Summarizes an S4 object}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{S4 object.}

\item{...}{Optional parameters.}
}
\value{
Summary.
}
\description{
Summarizes an S4 object.
}
\examples{
data("Amount")
abcResult = computeABCXYZAnalysis(data = Amount,
    value = "value",
    item = "item",
    timestamp = "date")
summary(abcResult)

}
\seealso{
\code{\link[=ABCXYZData-class]{summary}} \code{\link[=ABCXYZComparison-class]{summary}} \code{\link[=Forecast-class]{summary}}
}
\author{
Leon Binder \email{leon.binder@th-deg.de}

Bernhard Bauer \email{bernhard.bauer@th-deg.de}

Michael Scholz \email{michael.scholz@th-deg.de}
}
\keyword{methods}
