% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matman-package.R
\docType{package}
\name{matman-package}
\alias{matman-package}
\alias{matman}
\title{Material Management}
\description{
A set of functions, classes and methods for performing ABC and ABC/XYZ analyses, identifying overperforming, underperforming and
}
\details{
\tabular{ll}{ Package: \tab matman\cr Type: \tab Package\cr Version:
\tab 1.1.1\cr Date: \tab 2020-11-23\cr License: \tab GPL-3\cr Depends: \tab
R (>= 3.5.0), stats\cr }
}
\author{
Leon Binder \email{leon.binder@th-deg.de}

Bernhard Bauer \email{bernhard.bauer@th-deg.de}

Michael Scholz \email{michael.scholz@th-deg.de}
}
\concept{ABC-Analysis}
\concept{XYZ-Analysis}
